# target/generated-sources/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Documentación de los servicios Rest sobre HTTP expuestos por JBoss FUSE para manejo de pedidos, productos y recepciones del proyecto Customer Experience.",
    "version" : "1.0.0",
    "title" : "Newsan - Customer Experience",
    "contact" : {
      "name" : "Ingenia CA",
      "email" : "mail@ingenia.la"
    }
  },
  "basePath" : "/custexp",
...
  "schemes" : [ "https", "http" ],
  "paths" : {
...
    "/supplier/{category}" : {
      "get" : {
        "tags" : [ "suppliers" ],
        "summary" : "list EBS suppliers by vendor category filtered by last update date",
        "description" : "Consulta proveedores en EBS en base a la categoría y ultima fecha de actualización.",
        "operationId" : "listSuppliers",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "category",
          "in" : "path",
          "description" : "pass a supplier vendor category",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "lastUpdated",
          "in" : "query",
          "description" : "pass last updated date to filter products",
          "required" : true,
          "type" : "string",
          "pattern" : "\\d{4}-\\d{2}-\\d{2}"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful Operation",
            "schema" : {
              "type" : "object",
              "properties" : {
                "data" : {
                  "type" : "object",
                  "properties" : { }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid Parameters",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Supplier not Found",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/reception" : {
      "post" : {
        "tags" : [ "receptions" ],
        "summary" : "create a purchase order reception in EBS",
        "description" : "Crea una recepción de orden de compra en EBS",
        "operationId" : "createReception",
        "consumes" : [ "application/json" ],
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "in" : "body",
          "name" : "reception",
          "description" : "The reception to create",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/ApiReception"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful Operation",
            "schema" : {
              "type" : "object",
              "properties" : {
                "data" : {
                  "type" : "object",
                  "properties" : { }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid Parameters",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "type" : "object",
              "properties" : {
                "errors" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/definitions/Error"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions" : {
...
    "Error" : {
      "type" : "object",
      "properties" : {
        "code" : {
          "type" : "string"
        },
        "message" : {
          "type" : "string"
        }
      }
    },
...
    "ApiReception" : {
      "type" : "object",
      "required" : [ "createdBy", "lastUpdatedBy", "transactions", "vendor" ],
      "properties" : {
        "vendor" : {
          "type" : "string",
          "description" : "Identificación del proveedor"
        },
        "createdBy" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "Usuario de creación",
          "minimum" : -1,
          "default" : -1
        },
        "lastUpdatedBy" : {
          "type" : "integer",
          "format" : "int32",
          "description" : "Usuario de última actualización",
          "minimum" : -1,
          "default" : -1
        },
        "transactions" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ApiReceptionTransaction"
          }
        }
      }
    },
    "ApiReceptionTransaction" : {
      "type" : "object",
      "required" : [ "company", "purchaseOrder", "quantity", "sku", "uom" ],
      "properties" : {
        "processingMode" : {
          "type" : "string",
          "default" : "BATCH",
          "enum" : [ "ONLINE", "INMEDIATE", "BATCH" ]
        },
        "purchaseOrder" : {
          "type" : "integer",
          "description" : "Identificador orden de compra",
          "minimum" : 1
        },
        "company" : {
          "type" : "string",
          "description" : "Empresa donde se realiza la orden de compra"
        },
        "sku" : {
          "type" : "string",
          "description" : "Identificador del sku"
        },
        "quantity" : {
          "type" : "number",
          "format" : "float",
          "description" : "Cantidad de servicios a facturar",
          "minimum" : 1
        },
        "uom" : {
          "type" : "string",
          "description" : "Unidad de medida"
        }
      }
    }
  }
}